requirements.txt
cent==3.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==2.2.7
django-templated-mail==1.1.1
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
